Performance improvements

+++ src/shooter/gfx/GameCamera.java
@@ -11,8 +11,8 @@ public class GameCamera {
     }
 
     public void centerOnEntity(Entity e){               //hier wird die Welt soweit verschoben, dass es aussieht als würde sich der Viewport bewegen und nicht die Welt um den Spieler herum
-        xOffset = e.getX() - 960 + Entity.CREATURESIZE / 2;
-        yOffset = e.getY() - 540 + Entity.CREATURESIZE / 2;
+        xOffset = e.getX() - 960 + (float)Entity.CREATURESIZE / 2;
+        yOffset = e.getY() - 540 + (float)Entity.CREATURESIZE / 2;
     }
 
     public void move(float xAmt, float yAmt){           //mit move könnte man die Kamera bewegen, ohne dass sich der Spieler in der Mitte befindet
